requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
jmespath==1.0.1
lxml==5.2.2
openai==1.30.3
packaging==24.0
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.2
requests-file==2.1.0
resend==2.0.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
tldextract==5.1.2
tqdm==4.66.4
Twisted==24.3.0
typing_extensions==4.12.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.4.post2
